puchitto_project/requirements.txt
django
django-jazzmin
django-shortuuidfield
django-taggit
django-ckeditor
pillow